Do not free a bucket to be burst if burst did not succeed.

+++ htrie/htrie.h
@@ -2,7 +2,7 @@
  *		Tempesta DB
  *
  * Copyright (C) 2014 NatSys Lab. (info@natsys-lab.com).
- * Copyright (C) 2015-2022 Tempesta Technologies, Inc.
+ * Copyright (C) 2015-2023 Tempesta Technologies, Inc.
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by
@@ -47,7 +47,7 @@
 typedef struct {
 	union {
 		uint64_t 	col_map;
-		uint32_t	next;
+		uint64_t	next;
 	};
 } __attribute__((packed)) TdbHtrieBucket;
